+++ src/textual_textarea/__main__.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import sys
 
 from textual.app import App, ComposeResult
-from textual.widgets import Placeholder
+from textual.widgets import Footer, Placeholder
 
 from textual_textarea import TextEditor
 
@@ -13,6 +13,7 @@ class FocusablePlaceholder(Placeholder, can_focus=True):
 
 
 class TextApp(App, inherit_bindings=False):
+    BINDINGS = [("ctrl+q", "quit")]
     CSS = """
     TextEditor {
         height: 1fr;
@@ -26,20 +27,24 @@ class TextApp(App, inherit_bindings=False):
         try:
             language = sys.argv[1]
         except IndexError:
-            language = "python"
+            language = "sql"
         yield FocusablePlaceholder()
         self.editor = TextEditor(
             language=language,
-            theme="monokai",
             use_system_clipboard=True,
             id="ta",
         )
         yield self.editor
+        yield Footer()
+
+    def watch_theme(self, theme: str) -> None:
+        self.editor.theme = theme
 
     def on_mount(self) -> None:
+        self.theme = "gruvbox"
         self.editor.focus()
 
-        def _completer(prefix: str) -> list[tuple[str, str]]:
+        def _completer(prefix: str) -> list[tuple[tuple[str, str], str]]:
             words = [
                 "satisfy",
                 "season",
@@ -53,7 +58,7 @@ class TextApp(App, inherit_bindings=False):
                 "super",
                 "supercalifragilisticexpialadocioussupercalifragilisticexpialadocious",
             ]
-            return [(w, w) for w in words if w.startswith(prefix)]
+            return [((w, "word"), w) for w in words if w.startswith(prefix)]
 
         self.editor.word_completer = _completer