+++ tests/functional_tests/test_goto.py
@@ -22,6 +22,7 @@ async def test_goto_line(app: App) -> None:
         await pilot.press("2")
         await pilot.press("enter")
 
+        assert ta.text_input
         assert ta.text_input.has_focus
         assert ta.selection.start == ta.selection.end == (11, 0)